Major refactor of createWorkout and createExercise
