+++ frontend/kettlepal-fe/src/Components/NewWorkouts/FormComponents.tsx/Exercise/ExerciseTimer.tsx
@@ -1,0 +1,37 @@
+import { FormLabel, VStack } from "@chakra-ui/react";
+import Timer from "../Generic/Timer";
+import { CreateWorkoutState } from "../../CreateWorkout";
+
+interface ExerciseTimerProps {
+  exercise: Omit<CreateWorkoutState["exercises"][number], "key">;
+  timerIsActive: boolean;
+  handleTimerIsActive: (newState: boolean) => void;
+  setTime: (time: number) => void;
+}
+export default function ExerciseTimer({
+  exercise,
+  timerIsActive,
+  handleTimerIsActive,
+  setTime,
+}: ExerciseTimerProps) {
+  return (
+    <VStack
+      justifyContent={"flex-end"}
+      alignItems={"center"}
+      minWidth="130px"
+      spacing={0}
+    >
+      <FormLabel fontSize={["12px", "14px", "16px"]} m="0">
+        Elapsed Time
+      </FormLabel>
+      <Timer
+        seconds={exercise.elapsedSeconds}
+        isActive={timerIsActive}
+        handleIsActive={handleTimerIsActive}
+        setTime={setTime}
+        size="sm"
+        variant="digital"
+      />
+    </VStack>
+  );
+}
