+++ frontend/kettlepal-fe/src/Components/ViewWorkouts/ViewDetailedWorkoutModal/EditWorkout.tsx
@@ -1,0 +1,3 @@
+export default function EditWorkout() {
+  return <>{/* <WorkoutDate submitted={} state={} handleStateChange={}/> */}</>;
+}
